data_pipeline/requirements.txt
psycopg2==2.9.10
sqlalchemy==2.0.1
python-dotenv==1.0.1
kafka-python==2.1.2
bs4==0.0.2
schedule==1.2.2
pillow==11.1.0
numpy==2.2.4